requirements.txt
# Core image processing and machine learning libraries
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.4.0
scikit-image>=0.18.0

# Deep learning frameworks (choose one or both)
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

# Additional ML libraries
scipy>=1.7.0
scikit-learn>=1.0.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.62.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Optional: GUI libraries (uncomment if building a GUI)
# tkinter (usually comes with Python)
# PyQt5>=5.15.0
